requirements.txt
lightning==2.2.1
transformers==4.25.1
datasets==2.18.0
nltk==3.8.1
rouge-score==0.1.2
wandb==0.16.5
timm==0.5.4
yacs==0.1.8
rdkit==2023.09.6
gradio==5.12.0
gradio-client==1.5.4
torch-geometric==2.6.1
numpy==1.26.4
selfies==2.2.0
optuna
rapidfuzz